mcp/websearch-mcp/pyproject.toml
[project]
name = "websearch-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.2",
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    "python-dateutil>=2.9.0.post0",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
]
